+++ bank-feeds/src/funcs/accountMappingCreate.ts
@@ -131,7 +131,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/bankFeedAccounts/mapping",
   )(pathParams);