+++ bank-feeds/src/funcs/sourceAccountsGenerateCredentials.ts
@@ -106,7 +106,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/connectionInfo/bankFeedAccounts/credentials",
   )(pathParams);